+++ src/latch/__init__.py
@@ -20,6 +20,13 @@ from latch.resources.reference_workflow import workflow_reference
 from latch.resources.tasks import (
     custom_memory_optimized_task,
     custom_task,
+    g6e_xlarge_task,
+    g6e_2xlarge_task,
+    g6e_4xlarge_task,
+    g6e_8xlarge_task,
+    g6e_12xlarge_task,
+    g6e_16xlarge_task,
+    g6e_24xlarge_task,
     large_gpu_task,
     large_task,
     medium_task,