+++ src/latch/types/metadata/nextflow.py
@@ -130,10 +130,10 @@ class NextflowRuntimeResources:
     """
     Storage required for the task in GiB
     """
-    storage_expiration_hours: int = 0
+    storage_expiration_hours: int = 7 * 24
     """
     Number of hours after execution failure that workdir should be retained in EFS.
-    Warning: Increasing this number will increase your Nextflow EFS Storage costs.
+    Warning: Increasing this number will increase your Nextflow Storage costs.
     """